Updates to binutils patches
binutils 2.19.1, 2.20, and 2.20.1 were updated to correct the location of
the information memory and bootstrap loader files.

binutils 2.20 and 2.20.1 were updated to correct the references to
CC430x5123 which should have been CC430x5133.

Binutils patches are now generated from the binutils git repository hosted
on mspgcc4 with commands:

  for v in 2.19.1 2.20 2.20.1 ; do
    git diff binutils-${v}..patch/${v} > binutils-${v}.patch
  done

The resulting patches have a slightly different format than the previous
ones, but this should not affect their use.