SF 3055171: ICE in trunc_int_for_mode
msp430.md includes instructions that recognize values that can be generated
with the MSP430 constant generator (-1, 0, 1, 2, 4, and 8).  The
match_operand template to detect that in the case of addition to a register
did not include the required mode.  The represented value was a small
constant for which there is a shared generic RTL expression which is
intentionally marked as VOIDmode rather than any of the integral type
modes.  Consequently, without the mode marker on the match_operand, the
compiler was unable to ensure that the integral value could be truncated to
the appropriate size for the RTL operation.

This commit fixes only the specific instance tickled by the test case.  A
subsequent commit will update other rules that appear to have the same
problem.